PlotPriorAndPosterior {JASPgraphs}	R Documentation

Create a prior-posterior plot.

Description

Create a prior-posterior plot.

Usage

PlotPriorAndPosterior(dfLines, dfPoints = NULL, BF01 = NULL,
  CRI = NULL, median = NULL, xName = NULL, yName = "Density",
  drawPizzaTxt = !is.null(BF01), drawCRItxt = !is.null(CRI),
  bfSubscripts = "BF[0][1]", pizzaTxt = pizzaTxtFromBF(bfSubscripts),
  bty = list(type = "n", ldwX = 0.5, lwdY = 0.5), ...)

Arguments

dfLines
A dataframe with $x, $y, and optionally $g.

dfPoints
A dataframe with $x, $y, and optionally $g.

BF01
Numeric, with value of Bayes factor.

CRI
Numeric of length 2, Credible interval of posterior.

median
Numeric, median of posterior.

xName
String or expression, displayed on the x-axis.

yName
String or expression, displayed on the y-axis.

drawPizzaTxt
Logical, should there be text above and below the pizza plot?

drawCRItxt
Logical, should the credible interval be displayed in text?

bfSubscripts
String, to be parsed as expression. Indicates what BF type to display.

pizzaTxt
String vector of length 2, text to be drawn above and below pizza plot.

bty
List of three elements. Type specifies the box type, ldwX the width of the x-axis, lwdY the width of the y-axis.

...
Unused.

[Package JASPgraphs version 0.3.0 Index]