PlotRobustnessSequential {JASPgraphs}	R Documentation

Create a robustness or sequential plot

Description

Create a robustness or sequential plot

Usage

PlotRobustnessSequential(dfLines, dfPoints = NULL, BF01 = NULL,
  hasRightAxis = TRUE, xName = NULL, yName = bfSubscripts,
  addEvidenceArrowText = TRUE, drawPizzaTxt = !is.null(BF01),
  evidenceLeveltxt = !is.null(BF01), pointLegend = !is.null(dfPoints),
  linesLegend = !is.null(dfLines$g), bfSubscripts = "BF[0][1]",
  pizzaTxt = pizzaTxtFromBF(bfSubscripts), pointColors = c("grey",
  "white", "black", "red"), lineColors = c("black", "grey", "black"),
  lineTypes = c("dotted", "solid", "solid"), addLineAtOne = TRUE,
  bty = list(type = "n", ldwX = 0.5, lwdY = 0.5), ...)

Arguments

dfLines
A dataframe with $x, $y, and optionally $g.

dfPoints
A dataframe with $x, $y, and optionally $g.

BF01
Numeric, with value of Bayes factor.

hasRightAxis
Logical, should there be a right axis displaying evidence?

xName
String or expression, displayed on the x-axis.

yName
String or expression, displayed on the y-axis.

addEvidenceArrowText
Logical, should arrows indicating "Evidence for H0/H1" be drawn?

drawPizzaTxt
Logical, should there be text above and below the pizza plot?

evidenceLeveltxt
Logical, should "Evidence for H0: extreme" be drawn? Ignored if !is.null(dfLines$g) && linesLegend.

pointLegend
Logical, should a legend of dfPoints$g be shown?

linesLegend
Logical, should a legend of dfLines$g be shown?

bfSubscripts
String, to be parsed as expression. Indicates what BF type to display.

pizzaTxt
String vector of length 2, text to be drawn above and below pizza plot.

pointColors
String vector, colors for points if dfPoints$g is not NULL.

lineColors
String vector, colors for lines if dfLines$g is not NULL.

lineTypes
String vector, line types if dfLines$g is not NULL.

addLineAtOne
Logical, should a black line be draw at BF = 1?

bty
List of three elements. Type specifies the box type, ldwX the width of the x-axis, lwdY the width of the y-axis.

...
Unused.

[Package JASPgraphs version 0.3.0 Index]