JASPcolors {JASPgraphs}	R Documentation

JASP color palettes

Description

JASP color palettes

Usage


JASPcolors(n, palette = getGraphOption("palette")[["palette"]],
  asFunction = FALSE)


  scale_JASPcolor_continuous(palette = getGraphOption("palette")[["Function"]],
  ...)


  scale_JASPfill_continuous(palette = getGraphOption("palette")[["Function"]],
  ...)


  scale_JASPcolor_discrete(palette = getGraphOption("palette")[["Function"]],
  ...)


  scale_JASPfill_discrete(palette = getGraphOption("palette")[["Function"]],
  ...)

Arguments

n
Number of colors desired.

palette
Palette to choose from.

asFunction
Should a function be returned or the raw colors? If a function is returned, it either takes a single integer or a vector in 0, 1 as input.

...
Further arguments for ggplot2::scale_color_manual or ggplot2::scale_fill_manual.

Details

For ggplot2s, the convenience functions scale_JASPcolor_\* and scale_JASPfill_\* exist.

Value

Either a character vector of colors or a function.

Examples


library(ggplot2)

# use colors directly
colors <- JASPcolors(10, "colorblind")
plot(rnorm(10), col = colors, pch = 16, cex = 2)

colFun <- JASPcolors(palette = "colorblind", asFunction = TRUE)
plot(rnorm(10), col = colFun(10), pch = 16, cex = 2)

# use colorscales
e0 <- 1:10
df <- data.frame(x = e0, y = e0, z = e0, g = factor(e0))
ggplot(df, aes(x = x, y = y, color = z, fill = z)) +
  geom_point() +
  scale_JASPcolor_continuous() +
  scale_JASPfill_continuous()

ggplot(df, aes(x = x, y = y, color = g, fill = g)) +
  geom_point() +
  scale_JASPcolor_discrete() +
  scale_JASPfill_discrete()

[Package JASPgraphs version 0.3.7 Index]