drawBFpizza {JASPgraphs}	R Documentation

Make a pizza plot

Description

Make a pizza plot

Usage


drawBFpizza(dat, linewidth = 1, scaleText = 0.3, show.legend = FALSE,
  labels = NULL)

Arguments

dat
data.frame with $y, or vector.

linewidth
Scalar, linewidth of the black lines around the pizza.

scaleText
Scalar > 0, the size of the text is equal to scaleText * getGraphOption("fontsize").

show.legend
Logical, should a legend be shown?

labels
String vector of length 2, text to be shown above and below plot.

Examples


JASPgraphs::drawBFpizza(c(1, 4))

JASPgraphs::drawBFpizza(
  dat = data.frame(y = c(1, 4)),
 labels = c("data | H0", "data | H1")
)


[Package JASPgraphs version 0.3.0 Index]