parseThis {JASPgraphs}	R Documentation

Tell that this string should be parsed.

Description

Tell that this string should be parsed.

Usage

parseThis(x)

needsParsing(x)

Arguments

x
the object to be parsed, or the object to test if it will be parsed.

Details

parseThis adds an attribute to a character or factor such that JASPgraphs knows it should be parsed. needsParsing return TRUE if a character has said attribute.

[Package JASPgraphs version 0.3.0 Index]