Deprecated: use themeJaspRaw and geom_rangeframe instead.

Use themeJaspRaw and geom_rangeframe instead.

themeJasp(
  graph,
  xAxis = TRUE,
  yAxis = TRUE,
  sides = "bl",
  axis.title.cex = getGraphOption("axis.title.cex"),
  bty = getGraphOption("bty"),
  fontsize = getGraphOption("fontsize"),
  family = getGraphOption("family"),
  horizontal = FALSE,
  legend.position = getGraphOption("legend.position"),
  legend.justification = "top",
  axisTickLength = getGraphOption("axisTickLength"),
  axisTickWidth = getGraphOption("axisTickWidth")
)

Arguments

graph
a ggplot2 object

xAxis
Should there be a horizontal line for the x-axis?

yAxis
Should there be a horizontal line for the x-axis?

sides
see geom_rangeframe

axis.title.cex
scalar magnification for the title of the axes.

bty
remake R's bty = 'n'?

fontsize
global font size.

family
global font familiy.

horizontal
flip graph?

legend.position
where should the legend be?

legend.justification
see theme

axisTickLength
length of axis ticks.

axisTickWidth
width of axis ticks.