PlotRobustnessSequential {JASPgraphs}	R Documentation

Create a robustness or sequential plot

Description

Create a robustness or sequential plot

Usage


PlotRobustnessSequential(dfLines, dfPoints = NULL, BF = NULL,
  hasRightAxis = TRUE, xName = NULL, yName = NULL,
  addEvidenceArrowText = TRUE, drawPizzaTxt = !is.null(BF),
  evidenceLeveltxt = !is.null(BF), pointLegend = !is.null(dfPoints),
  linesLegend = !is.null(dfLines$g), bfSubscripts = NULL,
  pizzaTxt = hypothesis2BFtxt(hypothesis)$pizzaTxt, bfType = c("BF01",
  "BF10", "LogBF10"), hypothesis = c("equal", "smaller", "greater"),
  pointColors = c("red", "grey", "black", "white"),
  lineColors = c("black", "grey", "black"), lineTypes = c("dotted",
  "solid", "solid"), addLineAtOne = TRUE, bty = list(type = "n", ldwX =
  0.5, lwdY = 0.5), plotLineOrPoint = c("line", "point"),
  pointShapesSmallN = rep(21, 3), pointFillsSmallN = c("grey", "black",
  "white"), pointColorsSmallN = rep("black", 3), pointSizeSmallN = c(3,
  2, 2), ...)

Arguments

dfLines
A dataframe with $x, $y, and optionally $g. $y is assumed to be on the log scale.

dfPoints
A dataframe with $x, $y, and optionally $g.

BF
Numeric, with value of Bayes factor. This MUST correspond to bfType.

hasRightAxis
Logical, should there be a right axis displaying evidence?

xName
String or expression, displayed on the x-axis.

yName
String or expression, displayed on the y-axis.

addEvidenceArrowText
Logical, should arrows indicating "Evidence for H0/H1" be drawn?

drawPizzaTxt
Logical, should there be text above and below the pizza plot?

evidenceLeveltxt
Logical, should "Evidence for H0: extreme" be drawn? Ignored if !is.null(dfLines$g) && linesLegend.

pointLegend
Logical, should a legend of dfPoints$g be shown?

linesLegend
Logical, should a legend of dfLines$g be shown?

bfSubscripts
String, manually specify the BF labels.

pizzaTxt
String vector of length 2, text to be drawn above and below pizza plot.

bfType
String, what is the type of BF? Options are "BF01", "BF10", or "LogBF10".

hypothesis
String, what was the hypothesis? Options are "equal", "smaller", or "greater".

pointColors
String vector, colors for points if dfPoints$g is not NULL.

lineColors
String vector, colors for lines if dfLines$g is not NULL.

lineTypes
String vector, line types if dfLines$g is not NULL.

addLineAtOne
Logical, should a black line be draw at BF = 1?

bty
List of three elements. Type specifies the box type, ldwX the width of the x-axis, lwdY the width of the y-axis.

plotLineOrPoint
String, should the main geom in the plot be a line or a point?

pointShapesSmallN
String, if plotLineOrPoint == "point" then this controls the shape aestethic.

pointFillsSmallN
String, if plotLineOrPoint == "point" then this controls the fill aestethic.

pointColorsSmallN
String, if plotLineOrPoint == "point" then this controls the color aestethic.

...
Unused.

pointColorsSmallN
String, if plotLineOrPoint == "point" then this controls the size aestethic.

Examples


n <- 100
dfLines <- data.frame(
  x = seq_len(n),
  y = c(0, rnorm(n - 1, seq_len(n - 1) / 30, .5)) # log Bayes factor
)

PlotRobustnessSequential(
  dfLines      = dfLines,
  xName        = "n",
)

BF10 <- exp(tail(dfLines, 1)$y)
PlotRobustnessSequential(
  dfLines      = dfLines,
  xName        = "n",
  BF           = BF10,
  bfType       = "BF10"
)

[Package JASPgraphs version 0.3.7 Index]